+++ termal_core/src/raw/sys/unix.rs
@@ -1,18 +1,28 @@
 use std::{
-    fs, io, mem,
+    fs,
+    io::{self, ErrorKind},
+    mem,
     os::fd::{AsRawFd, IntoRawFd, RawFd},
+    ptr,
     sync::{Mutex, MutexGuard},
-    time::Duration,
+    time::{Duration, Instant},
 };
 
 use libc::{
-    EINTR, POLLIN, TCSANOW, TIOCGWINSZ, cfmakeraw, ioctl, poll, pollfd,
-    tcgetattr, tcsetattr, termios as Termios, winsize,
+    POLLERR, POLLIN, TCSANOW, TIOCGWINSZ, cfmakeraw, ioctl, poll, pollfd,
+    ppoll, tcgetattr, tcsetattr, termios as Termios, time_t, timespec,
+    winsize,
 };
 
-use crate::{Result, raw::TermSize};
+use crate::{Error, Result, raw::TermSize};
 
 static ORIGINAL_TERMINAL_MODE: Mutex<Option<Termios>> = Mutex::new(None);
+
+/// The maximum time you can wait for stdin. Larger values will return error.
+///
+/// This is unix implementation and thus the maximum is `time_t::MAX` seconds
+/// which is `i32::MAX` on 32 bit platforms and `i64::MAX` on 64 bit platforms.
+pub const MAX_STDIN_WAIT: Duration = Duration::from_secs(time_t::MAX as u64);
 
 fn get_original_terminal_mode() -> MutexGuard<'static, Option<Termios>> {
     ORIGINAL_TERMINAL_MODE
@@ -127,14 +137,81 @@ pub(crate) fn window_size() -> Result<TermSize> {
 /// Wait for stdin input on linux with the given timeout. If zero returns
 /// immidietly whether there is available input.
 pub(crate) fn wait_for_stdin(timeout: Duration) -> Result<bool> {
+    if timeout == Duration::MAX {
+        return infinite_wait_for_stdin();
+    }
+
+    if timeout > MAX_STDIN_WAIT {
+        return Err(Error::IntConvert);
+    }
+
+    let end = Instant::now() + timeout;
+
     let mut pdfs = pollfd {
         fd: libc::STDIN_FILENO,
         events: POLLIN,
         revents: 0,
     };
 
-    let r = unsafe { poll(&mut pdfs, 1, timeout.as_millis() as i32) };
-    Ok((r == 1 || r < 0) && r != EINTR)
+    loop {
+        let wait = end.saturating_duration_since(Instant::now());
+        let wait = timespec {
+            tv_sec: wait.as_secs() as time_t,
+            tv_nsec: wait.subsec_nanos().into(),
+        };
+        let r = unsafe { ppoll(&mut pdfs, 1, &wait, ptr::null()) };
+
+        match r {
+            0 => return Ok(false),
+            1 => {
+                return if (pdfs.revents & POLLERR) == POLLERR {
+                    Err(Error::WaitAbandoned)
+                } else {
+                    Ok(true)
+                };
+            }
+            -1 => {
+                let err = io::Error::last_os_error();
+                if err.kind() == ErrorKind::Interrupted {
+                    continue;
+                }
+                return Err(err.into());
+            }
+            _ => return Err(Error::WaitAbandoned),
+        }
+    }
+}
+
+fn infinite_wait_for_stdin() -> Result<bool> {
+    let mut pdfs = pollfd {
+        fd: libc::STDIN_FILENO,
+        events: POLLIN,
+        revents: 0,
+    };
+
+    loop {
+        let r = unsafe { poll(&mut pdfs, 1, -1) };
+
+        match r {
+            // Shouldn't happen
+            0 => return Ok(false),
+            1 => {
+                return if (pdfs.revents & POLLERR) == POLLERR {
+                    Err(Error::WaitAbandoned)
+                } else {
+                    Ok(true)
+                };
+            }
+            -1 => {
+                let err = io::Error::last_os_error();
+                if err.kind() == ErrorKind::Interrupted {
+                    continue;
+                }
+                return Err(err.into());
+            }
+            _ => return Err(Error::WaitAbandoned),
+        }
+    }
 }
 
 fn get_terminal_attr(fd: RawFd) -> Result<Termios> {